Create a mini gauge plot

This function creates a mini gauge plot using the plotly package. The gauge represents a value within a range, divided into color-coded steps and highlighting a threshold value.

Usage

minigauge(
  value,
  breaks = c(0.15, 0.3),
  colors = c("red", "#f7cf07", "#3af72c")
)

Arguments

value

numeric value to be represented on the gauge (must be between 0 and 1)

breaks

numeric vector specifying the break points for coloring the gauge steps (must be two elements)

colors

character vector specifying the colors for each step in the gauge (must be three elements)

Value

a plotly object representing the mini gauge plot

Examples

minigauge(0.25)